Refactor: organisms/category 디렉토리 제거,사용하지 않는 파일 삭제
