--- src/components/organisms/category/index.js
@@ -1,3 +1,0 @@
-export { default as Advertisement } from './Advertisement';
-export { default as LeftBar } from './LeftBar';
-export { default as PageSelector } from '../../../common/PageSelector';
